chore: rename Codec -> Encoding
