+++ src/MaaCore/Utils/Encoding.cpp
@@ -1,4 +1,4 @@
-#include "Utils/Codec.h"
+#include "Utils/Encoding.h"
 
 #include <cstdint>
 #include <string>
