+++ src/MaaCore/Utils/Encoding.h
